Refactor various components to ensure proper type casting and improve code clarity
- Updated rowCount and count methods in ChatMessageListModel and AgentActivityLog to use static_cast for size conversions.
- Introduced a utility function toInt for consistent conversion from qsizetype to int across multiple files.
- Enhanced the applyVoiceConfigurationPaths method in DesktopShellViewModel with comments for clarity.
- Added a None value to the RetrievalSourcePriority enum for better handling of retrieval priorities.
- Refactored InMemoryConversationStore and InMemoryMemoryCandidateStore to ensure proper size handling.
- Adjusted InMemorySettingsStore, InMemoryStore, and InMemoryToolRegistry to avoid unnecessary move semantics.
- Improved test cases in InMemoryToolRegistryTest for better error handling and clarity.

+++ core/src/AgentActivityLog.cpp
@@ -25,7 +25,7 @@ void AgentActivityLog::clear() {
 }
 
 int AgentActivityLog::count() const {
-    return entries_.size();
+    return static_cast<int>(entries_.size());
 }
 
 QString AgentActivityLog::latestSummary() const {
